Add gfx interface definition in Slang. (#2364)
* Add gfx interface definition in Slang.

- add gfx interface definitons in Slang.
- fix slang compiler to correctly type-check `out` interface argument.
- modify gfx interface to be fully COM compatible
- add convenient ShaderProgram creation methods to gfx.

* Fix compile errors and warnings.

* Update project files

* Fix cuda.

* Properly implement queryInterface in command encoder impls.

+++ tools/gfx-unit-test/gfx-test-util.cpp
@@ -151,7 +151,7 @@ namespace gfx_test
 
     void compareComputeResultFuzzy(const float* result, float* expectedResult, size_t expectedBufferSize)
     {
-        for (int i = 0; i < expectedBufferSize / sizeof(float); ++i)
+        for (size_t i = 0; i < expectedBufferSize / sizeof(float); ++i)
         {
             SLANG_CHECK(abs(result[i] - expectedResult[i]) <= 0.01);
         }
